# add new interface vaMapBuffer2 for map operation optimization