@IlyaSergeyev 2020-07-30
@XinfengZhang What the current status? This change can significantly improve encoding performance.