@nowrep 2023-10-07
What's the status of this? It would be great to have it merged.
Currently the issue in Mesa is that mapping all buffers read+write leads to bad performance on dGPUs.