@nowrep 2023-10-07
That sounds good, thanks!

My interest for this PR is radeonsi but it will benefit all Mesa drivers, so that also includes vaon12.