@dvrogozh 2023-10-07
Is there any discussion on mesa side regarding this? Maybe some issue filed?