@nowrep 2023-10-09
Mesa implementation: https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/25620