@dvrogozh 2023-10-10
Thank you @nowrep for quick follow up. Quoting perf example from https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/25620#note_2119488:

"Example with patched ffmpeg to use vaMapBuffer2:
```
ffmpeg -vaapi_device /dev/dri/renderD128 -f lavfi -i testsrc=size=1280x720 -vf format=nv12,hwupload -c:v h264_vaapi -async_depth 1 -f null -

Before (read+write): 540fps
After  (write only): 620fps
```

It's around 10-15% faster depending on the resolution. At 2048x2048 it's 160fps before and 180fps after."